requirements.txt
matplotlib
numpy
scipy
sklearn
pillow